Startup refactor
Changing how combat startup is handled.

+++ Assets/Scripts/Combat_Scene/Tiles/Tile_Burst.cs
@@ -10,7 +10,7 @@ public class Tile_Burst : MonoBehaviour
     // The object's particle system to release sparkles
     private ParticleSystem PS;
 
-    public void Activate(TColor _tint, int _blips, int _pointVal, int _reviveVal)
+    public void Activate(TColor _tint, int _blips, float _pointVal, float _reviveVal)
     {
         PS = GetComponent<ParticleSystem>();
         ParticleSystem.MainModule tempMain = PS.main;
@@ -23,7 +23,7 @@ public class Tile_Burst : MonoBehaviour
             {
                 bool isDead = GameManager.instance.party.GetPlayer(_tint).Status.IsDead;
                 Transform barPos = isDead ? Combat_UI_Commands.GetReviveBarPos((int)_tint) : Combat_UI_Commands.GetEnergyBarPos(_tint);
-                int potency = isDead ? _reviveVal : _pointVal;
+                int potency = (int)(isDead ? _reviveVal : _pointVal);
                 
                 int remainder = potency % _blips;
                 int blipPotency = Mathf.FloorToInt(potency / _blips);
@@ -43,7 +43,7 @@ public class Tile_Burst : MonoBehaviour
                 {
                     bool isDead = GameManager.instance.party.GetPlayer(i).Status.IsDead;
                     Transform barPos = isDead ? Combat_UI_Commands.GetReviveBarPos(i) : Combat_UI_Commands.GetHealthBarPos(i);
-                    int potency = isDead ? _reviveVal : _pointVal;
+                    int potency = (int)(isDead ? _reviveVal : _pointVal);
 
                     GameObject tempBlip = Instantiate(normalBlip, transform.position, Quaternion.identity);
                     tempBlip.GetComponent<Energy_Blip>().Activate(_tint, i, barPos, potency);
@@ -56,7 +56,7 @@ public class Tile_Burst : MonoBehaviour
                 {
                     bool isDead = GameManager.instance.party.GetPlayer(i).Status.IsDead;
                     Transform barPos = isDead ? Combat_UI_Commands.GetReviveBarPos(i) : Combat_UI_Commands.GetEnergyBarPos(i);
-                    int potency = isDead ? _reviveVal : _pointVal;
+                    int potency = (int)(isDead ? _reviveVal : _pointVal);
 
                     GameObject tempBlip = Instantiate(normalBlip, transform.position, Quaternion.identity);
                     tempBlip.GetComponent<Energy_Blip>().Activate(_tint, i, barPos, potency);
